Let's Encrypt certbot-auto 중복 설치 방지

+++ scripts/letsencrypt-install.sh
@@ -8,22 +8,26 @@ source "${STACK_ROOT}/includes/function.inc.sh"
 
 title "Let's Encrypt 자동화툴을 설치합니다."
 
-if [ -f "/usr/bin/certbot-auto" ]; then
-  abort "이미 /usr/bin/certbot-auto 이 설치되어 있으므로 중단합니다.";
-fi
-
 notice "[1/5] 인증툴인 certbot-auto 를 설치합니다. - https://certbot.eff.org/"
-cd /usr/bin \
-&& wget https://dl.eff.org/certbot-auto \
-&& chmod ug+x certbot-auto \
-&& certbot-auto --install-only -n
+if [ ! -f "/usr/bin/certbot-auto" ]; then
+  cd /usr/bin \
+  && wget https://dl.eff.org/certbot-auto \
+  && chmod ug+x certbot-auto \
+  && certbot-auto --install-only -n
+else
+  outputComment "  (생략) /usr/bin/certbot-auto 파일이 이미 존재합니다.\n"
+fi
 
 notice "[2/5] nginx 에서 사용할 설정 파일들을 /etc/letsencrypt/php79 에 복사합니다."
-if [ ! -d /etc/letsencrypt ]; then
-  mkdir /etc/letsencrypt
-fi
-if [ ! -d /etc/letsencrypt/php79 ]; then
-  cp -av "${STACK_ROOT}/letsencrypt/php79" /etc/letsencrypt/
+if [ ! -f "/etc/letsencrypt/php79/ssl-stable.conf" ]; then
+  if [ ! -d /etc/letsencrypt ]; then
+    mkdir /etc/letsencrypt
+  fi
+  if [ ! -d /etc/letsencrypt/php79 ]; then
+    cp -av "${STACK_ROOT}/letsencrypt/php79" /etc/letsencrypt/
+  fi
+else
+  outputComment "  (생략) /etc/letsencrypt/php79/ssl-stable.conf 파일이 이미 존재합니다.\n"
 fi
 
 notice "[3/5] 인증용 웹문서 디렉토리를 생성합니다."
@@ -32,8 +36,16 @@ if [ ! -d /var/www/letsencrypt/.well-known ]; then
 fi
 
 notice "[4/5] 암호화 강화를 위한 Diffie-Hellman parameters 를 생성합니다."
-openssl dhparam -out /etc/letsencrypt/php79/ssl-dhparams.pem 2048
+if [ ! -f "/etc/letsencrypt/php79/ssl-dhparams.pem" ]; then
+  openssl dhparam -out /etc/letsencrypt/php79/ssl-dhparams.pem 2048
+else
+  outputComment "  (생략) /etc/letsencrypt/php79/ssl-dhparams.pem 파일이 이미 존재합니다.\n"
+fi
 
 notice "[5/5] 인증서 갱신 스크립트를 매일 새벽 자동 실행되도록 크론에 추가합니다."
-cp -av "${STACK_ROOT}/letsencrypt/php79-certbot-auto-renew" /etc/cron.daily/
-chmod 700 /etc/cron.daily/php79-certbot-auto-renew
+if [ ! -f "/etc/cron.daily/php79-certbot-auto-renew" ]; then
+  cp -av "${STACK_ROOT}/letsencrypt/php79-certbot-auto-renew" /etc/cron.daily/
+  chmod 700 /etc/cron.daily/php79-certbot-auto-renew
+else
+  outputComment "  (생략) /etc/cron.daily/php79-certbot-auto-renew 파일이 이미 존재합니다.\n"
+fi
